# Faster rem_pio2 kernel for `Float32`