This has been tested to give the correct answer on every `Float32` and makes `sin(::Float32)` roughly .5ns faster.

Before:
```julia> x32 = 16 *rand(Float32, 1024); y32 = similar(x32);

julia> @benchmark @. $y32 = sin($x32)
BenchmarkTools.Trial: 10000 samples with 7 evaluations.
 Range (min … max):  4.252 μs …   9.946 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     4.428 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   4.478 μs ± 225.102 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

    ▁▆▄█▆▅▃▁▂ ▂ ▁                                              
  ▃▇██████████████▇▇▇▇▆▆▅▅▄▄▄▄▄▃▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂ ▄
  4.25 μs         Histogram: frequency by time        5.25 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> x32 = 2^20 *rand(Float32, 1024); y32 = similar(x32);

julia> @benchmark @. $y32 = sin($x32)
BenchmarkTools.Trial: 10000 samples with 7 evaluations.
 Range (min … max):  4.697 μs …  37.738 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     4.929 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   4.951 μs ± 615.364 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █ █▆ ▆ ▅  ▂▇ █ ▄▄ ▄ ▁▅▁▄▃ ▂  ▁  ▁▂ ▂                        ▂
  █▃██▂█▂██▃████▇██▅█▆████████████████▇▇▇▇▆▆▇▆▆▆▇▆▃▆▅▅▂▂▅▅▅▄▄ █
  4.7 μs       Histogram: log(frequency) by time      5.93 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.
```
After:
```
julia> x32 = 16 *rand(Float32, 1024); y32 = similar(x32);

julia> @benchmark @. $y32 = sin($x32)
BenchmarkTools.Trial: 10000 samples with 8 evaluations.
 Range (min … max):  3.666 μs …   9.352 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     3.752 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   3.853 μs ± 445.950 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ██ ▅ ▁▆▄▁                                                   ▂
  ██████████▇▆▆▆▅▅▄▆▄▆▅▅▁▅▅▃▄▄▄▃▃▁▃▁▅▄▁▃▃▁▃▁▁▃▄▃▄▅▄▇▅▅▅▅▄▄▅▄▆ █
  3.67 μs      Histogram: log(frequency) by time      6.64 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> x32 = 2^20 *rand(Float32, 1024); y32 = similar(x32);

julia> @benchmark @. $y32 = sin($x32)
BenchmarkTools.Trial: 10000 samples with 8 evaluations.
 Range (min … max):  3.759 μs …  14.508 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     3.784 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   3.863 μs ± 289.041 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ██  ▁█▂  ▂          ▃▄                                      ▂
  ██▇▃███▆▇███████▇▇▇▇██▇▇█▇▇█▇▇▇▆▆▆▅▄▄▅▆▅▁▄▃▄▄▃▅▄▃▃▅▃▅▃▄▃▄▄▄ █
  3.76 μs      Histogram: log(frequency) by time      4.79 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.
```